Refactor container model

+++ lib/business/rebuild_container.rb
@@ -1,13 +1,16 @@
 module Business
 	class RebuildContainer
-    attr_reader :options, :recommended_image
+    attr_reader :options, :recommended_image, :to_be_deleted_container
     def initialize(options = {container_id: "0efettttttt"})
       @options = options
     end
 
     def execute
+      @to_be_deleted_container = Container.to_be_deleted_container(options[:container_id])
+      return "[warning] The container to be deleted doesn't exist." unless to_be_deleted_container
       to_be_deleted_container.update_attributes(status: Container::STATUS_LIST['deleted'])
       @recommended_image = Image.recommended_image(to_be_deleted_container.image.purpose)
+      return "[warning] There's none recommended image for #{to_be_deleted_container.image.purpose} purpose." unless recommended_image
       begin
         request = Service::Docker::Request.new(docker_remote_api: to_be_deleted_container.ip_address.device.docker_remote_api)
         request.delete_container(container: to_be_deleted_container.container_id)
@@ -17,19 +20,10 @@ module Business
         @container_id = result.to_hash["Id"]
         request.start_container(container: @container_id)
         create_container_record
-        puts @container_id
+        @container_id
       rescue => e
 
       end
-    end
-
-    def to_be_deleted_container
-      @to_be_deleted_container ||=
-        begin
-          Container.where(container_id: options[:container_id]).first
-        rescue
-          nil
-        end
     end
 
     def create_container_record
